src/api_admin/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
boto3==1.28.62
botocore==1.31.62
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
joblib==1.3.2
lightgbm==4.0.0
mysql-connector-python==8.1.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pluggy==1.3.0
protobuf==4.21.12
pydantic==2.3.0
pydantic_core==2.6.3
pytest==7.4.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.31.0
s3transfer==0.7.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
